20250519 继续重构以减少项目整体的耦合度
名称标准化&增加注释：
Archive
Audio
Concurrent
CLI
Crypto KDF&Hash
SQL DAO Generator
ASM & ASMX
修正@since格式

change readme & doc

HTTP
修复FileContent处理零字节文件的bug
修复Request#getURL函数的问题
为FileShare增加本地和远程文件数量限制

Library
修改7z压缩包处理，现在能正确处理属性乱序的情况了，同时还支持流式读取，减少内存峰值
2-pass compression

Config
减小Tokenizer的String生成数量
加入DoubleParser

CLI
修复ASAR解包器的bug

Crypto
加入L64W64X128MixRandom & WYHash / WYRand

Audio
加入FFTAudioSink
Might OK Audio codec

+++ java/roj/archive/zip/EntryMod.java
@@ -4,7 +4,7 @@ import org.intellij.lang.annotations.MagicConstant;
 
 /**
  * @author Roj234
- * @since 2023/3/14 0014 0:42
+ * @since 2023/3/14 0:42
  */
 public final class EntryMod {
 	String name;
@@ -18,7 +18,7 @@ public final class EntryMod {
 
 	public byte[] pass;
 	@MagicConstant(intValues = {ZipArchive.CRYPT_NONE, ZipArchive.CRYPT_ZIP2, ZipArchive.CRYPT_AES, ZipArchive.CRYPT_AES2})
-	public int cryptType;
+	public byte cryptType;
 
 	public String getName() { return name; }
